Merkle refactoring
